check right to edit or delete

+++ investment/views.py
@@ -108,6 +108,8 @@ def detail(request, category_slug, post_id):
 """
 
 def add(request):
+    if not request.user.is_authenticated:
+        return redirect("study:index")
     # 템플릿 지정
     template_name = "investment/investment_add.html"
 
@@ -129,6 +131,8 @@ def add(request):
     return render(request, template_name, context)
 
 def edit(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect("study:index")
     # 수정할 포스트 획득
     post = get_object_or_404(Post, pk=post_id)
 
@@ -157,6 +161,8 @@ def edit(request, post_id):
 
 
 def delete(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect("study:index")
     post = get_object_or_404(Post, pk=post_id)
     post.delete()
     return redirect("investment:index")
